Upstream version 24.0~b3

# gpg: Signature made Mo 16 Sep 2013 18:19:23 CEST
# gpg:                using DSA key 0x9FCF2CCD3F3E6426
# gpg: please do a --check-trustdb
# gpg: Good signature from "Guido Günther <agx@sigxcpu.org>"
# gpg:                 aka "Guido Guenther <agx@debian.org>"
# gpg:                 aka "Guido Guenther <agx@sigxcpu.org>"

+++ mailnews/base/util/nsMsgProtocol.h
@@ -110,7 +110,7 @@ protected:
   // As data arrives on the socket, OnDataAvailable calls ProcessProtocolState.
   
   virtual nsresult ProcessProtocolState(nsIURI * url, nsIInputStream * inputStream, 
-                  uint32_t sourceOffset, uint32_t length) = 0;
+                                        uint64_t sourceOffset, uint32_t length) = 0;
 
   // SendData -- Writes the data contained in dataBuffer into the current output stream. 
   // It also informs the transport layer that this data is now available for transmission.
@@ -158,7 +158,7 @@ protected:
   nsCOMPtr<nsIInterfaceRequestor> mCallbacks;
   nsCOMPtr<nsISupports>       mOwner;
   nsCString                   m_ContentType;
-  int32_t                     mContentLength;
+  int64_t                     mContentLength;
 
   nsCString m_lastPasswordSent; // used to prefill the password prompt
